refactor: Split user/penpal names into separate first/last name fields
This is to ensure that last names don't get leaked to people they
shouldn't - penpals shouldn't be able to see each other's last names,
but admins should. This is done by removing the :name field from the
database for both users and penpals, and adding new :first_name and
:last_name fields.

The migration for this DOES NOT decrypt/re-encrypt names, and as such,
after the migration is applied all name data will be lost and everyone
will have to set their names again. This is because connecting to
keyderiv and loading in the crypto code isn't feasible while performing
migrations - it could be possible when called via `rake db:migrate` but
would break entirely if using the Sequel CLI.

The Penpal#get_name method has been modified to return a [:first_name,
:last_name] array, and a User#get_name method has been added that does
the same. All uses of Penpal#get_name have been changed to accommodate
this - which involves flattening this into a string whenever it's shown
to admins, and grabbing the first item in the array (so, the first name)
when showing this to end users. I'm pretty confident I've caught every
use, but if you notice something, file an issue (or MR). In any case
where the change hasn't been applied, you'll see `["First", "Last"]` in
the UI instead of a flattened string (which isn't all that bad, as it
still works, it's just ugly).