Add support for an "administration profile"
The administration profile is a penpal profile with no associated user
but without the "is incarcerated" flag set. Any uploaded correspondence
directed to this penpal sends an email notification to the site alert
email defined in the 'site-alert-emails' configuration entry, and those
email notifications use a different template from other notification
emails.

The ID of this profile should be set in the 'admin-profile-id'
configuration entry, and if this entry is set (and is a valid penpal
ID), any new penpals created via `/system/penpal/create` automatically
have a relationship with the admin profile created.

The `CreateAdministratorProfileRelationshipWorker` background worker
creates relationships between all incarcerated penpals and the set
administration profile. This worker should be run after creating an
administration profile.

Currently, the administration profile has to be manually created and the
ID set in the configuration, however this will change soon.