system/config: Always do a dry run of refresh before applying changes
This changes /system/configuration/refresh to accept GET, which will do
a dry run of the refresh, presenting any warnings as well as a button to
initate a real refresh. This reduces the chance of parse errors screwing
up the application and causing 5xx errors for end users if a config
change is accidentally broken.