app: Add user groups
This adds the models Group, GroupRole, and UserGroup. A UserGroup is the
connection between a User and a Group, and GroupRole is a role that is
applied to all members of the group (this is implemented in
`has_role?`).

This implementation does not yet have a way to create or edit groups.