system/penpal: Allow overriding automatic penpal status modification
This adds a checkbox to the penpal edit screen that sets the flag
`status_override` on the penpal object. If this flag is set, automatic
status changes don't happen for that penpal. This is accomplished by
pushing `false` to the `do_transition` list at the start of the penpal
handling in PenpalStatusTransitionWorker.