auth/login: Move TOTP MFA to own controller, merge recovery code handling
This starts the process of renaming "two-factor" to "multi-factor".

This also integrates the recovery code handling into the same form,
allowing typing a recovery code into the TOTP field, which brings it
in line with how EARMMS currently handles this. Currently, any input
over 8 characters (after removing spaces and hyphens) is assumed to be
a recovery code.