system/roster: Add volunteer roster (#17)
This allows users with role `system:roster:access` to view the roster,
and allows users with `system:roster:self_assign` to assign themselves
to a day on the roster.

Users with `system:roster:admin` are able to override roster entries -
that is, they can delete a roster entry for a given day (allowing it
to be self-assigned by another volunteer), or pick a user to assign to
a given roster day.

There's no email notifications for rostered users, yet - that will be
implemented soon.