auth/signup: verify email before account creation
This commit splits the signup process in two (the clear divide of which
can be seen by looking at the diff!), making it so that signups require
verifying the email address of the user.

On signup, the user's information is now stored as encrypted JSON in the
extra_data field of a token in the database, and an account is not
actually created. An email is sent to the user's email address,
containing a link to "verify [their] account," which contains the
aforementioned token in the URL.

Clicking that link (`/auth/signup/verify/TOKEN`) will actually create
the account using the data stored against the token, send the welcome
email and the admin alert for the new signup, and log the new user in.