user: add "soft deletion"
This adds the `User#soft_deleted` flag - if set, everything user-facing
acts like the user doesn't exist (including not sending emails), but the
user still exists within the system (and the user's correspondence is
kept).

Also adds the `User#purge_at_next_opportunity` flag - if set, the user
will be permanently deleted at the next run of the AccountPurgeWorker,
which is set to run every 8 hours.