compose: require entering pseudonym to sign letter with
This adds a required "pseudonym" field underneath the correspondence
composer, which is pre-populated with the user's first name (or, if one
has been defined, their pseudonym). The pseudonym provided in this field
is automatically added to the end of the outgoing correspondence, aiming
to solve the issue of people forgetting to sign their names on outgoing
correspondence.

See againstprisons/reconnect#20 for the reasoning for this change.