Holiday card implementation: part 1
This allows configuring holiday card "instances" (for a given event),
adds a background worker to create relationships and count entries for
each penpal matching the instance conditions, and adds a helper for
retrieving a list of penpals to present to the user to write to.

This is all backend stuff, there's no user interface to any of it atm.