Holiday card implementation: part 2
This adds a partial user interface for holiday card instances. It
currently only supports the "manual" count update mode, actually
writing holiday cards is not yet implemented.

The "manual" count update mode allows updating the :manual_count field
of HolidayCardCount for the current instance, for the list of penpals
that are displayed on the page.

The penpals displayed on the page are randomly selected from the list
of HolidayCardCount entries for the given instance, where preference is
given to penpals with a lower count for the selected count field (which
is either `:manual_count` or `:online_count`, selectable by the option
`:count_field` to HolidayCardHelpers#holidaycard_penpal_list).

Part 3 will contain the start of the implementation of the online card
writing process.